fetch/metadata/generated/font-sec-fetch-dest.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-Dest</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.js"></script>
  <body>
  <div id="font-test">test</div>
  <script>
  'use strict';

  function induceRequest(t, path) {
    const style = document.createElement('style');
    style.innerText = `
      @font-face {
        font-family: myWptFont;
        src: url(${path});
      }
      #font-test {
        font-family: myWptFont;
      }
    `;

    document.body.appendChild(style);

    t.add_cleanup(() => style.remove());
  }

  promise_test((t) => {
    const id = '{{uuid()}}';
    const path = '/fetch/metadata/resources/record-headers.py?key=' + id;

    induceRequest(t, '' + path);

    return pollForRequestRecording(id)
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_equals(headers['sec-fetch-dest'], 'font');
        });
  }, 'sec-fetch-dest - ');
  </script>
  </body>
</html>
